[ADD]requestのロジックをディレクトリ分離する
