+++ src/views/sidebar.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import axios, { AxiosRequestConfig, AxiosResponse, AxiosError } from "axios";
+import { requestDiscordWebhook } from "../services/request";
 
 // TODO: anyの排除
 export class SideBarView {
@@ -21,18 +21,7 @@ export class SideBarView {
   }
 
   notifyDiscord() {
-    const options: AxiosRequestConfig = {
-      method: "POST",
-      url: "https://discord.com/api/webhooks/989950740145516595/PDgWyUkErpiWXzJ7DHG8y8i4nyYzt9Xl_gj5fMEcNcQQ-lrV0WmM3x_qHtZyLcF4l-qk",
-      data: {
-        content: "Hello World!",
-      },
-      headers: {
-        "Content-Type": "application/json",
-      },
-    };
-
-    axios(options);
+    requestDiscordWebhook();
 
     vscode.window.showInformationMessage("稼働報告しました");
   }
